split section and information pages into components so they can share a route
